:recycle: Refactor express route methods into own helper file
