+++ lib/cards/express-route-alert.js
@@ -1,10 +1,10 @@
+const { assembleAdaptiveCard, localizeDateTime } = require('../helpers');
+
 const {
-  assembleAdaptiveCard,
   initPeersStatus,
   setPeerStatus,
   whichPeer,
-  localizeDateTime,
-} = require('../helpers');
+} = require('../express-route');
 
 exports.messageCard = async (alertData) => {
   await initPeersStatus();
